Ordered bytes in rawDigest output
Rusha#rawDigest used to return an Int32Array with the byte order
of the underlying platform and not sliced, which made the usage of the
underlying ArrayBuffer obtained by Int32Array#buffer hard (one had to
copy in another array using the proper endianness).

Changes a bit the API as rawDigest now reportedly returns an
ArrayBuffer instead of an Int32Array. This ArrayBuffer is sliced
at 20 bytes.

Added tests using the string 'abc' as an example.